Capture literature, and capture runner
capture.py will either display options or dispatch a specific capture
script.

The literature capture will prompt for elements of a new academic
article, and also fetch the bibtex entry if a DOI is given. It will then
fill the clipboard in 2 stages (1: entry, for pasting into literature
file; 2: filename, for renaming the downloaded file)