Update nasutil
Now, commands are implemented as methods on a struct that holds the list
of urls to download and urls that have failed. This lets me load all
data at the beginning of the program, and write all data at the end of
the program, without manually juggling two lists and two mildly
different save methods.